test: extract tool orchestrator white-box suites
