+++ crates/agent-engine/src/runtime/tool_orchestrator.rs
@@ -1096,2196 +1096,4 @@ where
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        agent_machine::AgentMachine,
-        config::Config,
-        runtime::{NamespaceRuntimeEnvironment, TurnState},
-        tools::{Tool, ToolContext, ToolRegistry, ToolResult},
-    };
-    use alan_agent_protocol::ToolCapability;
-    use alan_agentfs::AgentFs;
-    use alan_ap::{
-        ErrorCode, Fid, FileKind, FileServer, InProcessTransport, OpenMode, Qid, Request, Stat,
-    };
-    use alan_kernel::{
-        Access, Credentials, MountFs, Namespace, ProcFs, ProcessInvocation, ProcessOutcome,
-        ProcessRunner,
-    };
-    use alan_llm::{GenerationRequest, GenerationResponse, LlmProvider, StreamChunk};
-    use alan_shell::Shell;
-    use async_trait::async_trait;
-    use std::collections::HashMap;
-    use std::path::PathBuf;
-    use std::sync::{
-        Arc,
-        atomic::{AtomicUsize, Ordering},
-    };
-    use tokio::sync::Mutex;
-
-    // Simple mock provider for testing
-    struct SimpleMockProvider;
-
-    #[async_trait]
-    impl LlmProvider for SimpleMockProvider {
-        async fn generate(
-            &mut self,
-            _request: GenerationRequest,
-        ) -> anyhow::Result<GenerationResponse> {
-            Ok(GenerationResponse {
-                content: "test".to_string(),
-                thinking: None,
-                thinking_signature: None,
-                redacted_thinking: Vec::new(),
-                tool_calls: vec![],
-                usage: None,
-                finish_reason: None,
-                warnings: Vec::new(),
-                provider_response_id: None,
-                provider_response_status: None,
-            })
-        }
-
-        async fn chat(&mut self, _system: Option<&str>, _user: &str) -> anyhow::Result<String> {
-            Ok("mock".to_string())
-        }
-
-        async fn generate_stream(
-            &mut self,
-            _request: GenerationRequest,
-        ) -> anyhow::Result<tokio::sync::mpsc::Receiver<StreamChunk>> {
-            let (tx, rx) = tokio::sync::mpsc::channel(1);
-            let _ = tx
-                .send(StreamChunk {
-                    text: Some("test".to_string()),
-                    thinking: None,
-                    thinking_signature: None,
-                    redacted_thinking: None,
-                    usage: None,
-                    provider_response_id: None,
-                    provider_response_status: None,
-                    sequence_number: None,
-                    tool_call_delta: None,
-                    is_finished: true,
-                    finish_reason: Some("stop".to_string()),
-                })
-                .await;
-            Ok(rx)
-        }
-
-        fn provider_name(&self) -> &'static str {
-            "mock"
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    enum BinNode {
-        Root,
-        Tool(&'static str),
-    }
-
-    const BUILTIN_BIN_TOOLS: [&str; 7] = [
-        "read_file",
-        "write_file",
-        "edit_file",
-        "bash",
-        "grep",
-        "glob",
-        "list_dir",
-    ];
-
-    struct StaticBinFs {
-        fids: Mutex<HashMap<Fid, BinNode>>,
-    }
-
-    impl StaticBinFs {
-        fn new() -> Self {
-            let mut fids = HashMap::new();
-            fids.insert(Fid::ROOT, BinNode::Root);
-            Self {
-                fids: Mutex::new(fids),
-            }
-        }
-    }
-
-    #[async_trait]
-    impl FileServer for StaticBinFs {
-        async fn walk(&self, fid: Fid, newfid: Fid, names: &[String]) -> Result<Qid, ErrorCode> {
-            let mut fids = self.fids.lock().await;
-            if newfid == Fid::ROOT || fids.contains_key(&newfid) {
-                return Err(ErrorCode::BadRequest);
-            }
-            let mut node = *fids.get(&fid).ok_or(ErrorCode::NotFound)?;
-            for name in names {
-                node = match (node, name.as_str()) {
-                    (BinNode::Root, name) => BUILTIN_BIN_TOOLS
-                        .iter()
-                        .copied()
-                        .find(|tool| *tool == name)
-                        .map(BinNode::Tool)
-                        .ok_or(ErrorCode::NotFound)?,
-                    (BinNode::Tool(_), _) => return Err(ErrorCode::NotDirectory),
-                };
-            }
-            fids.insert(newfid, node);
-            Ok(bin_qid(node))
-        }
-
-        async fn open(&self, fid: Fid, mode: OpenMode) -> Result<Qid, ErrorCode> {
-            let fids = self.fids.lock().await;
-            let node = *fids.get(&fid).ok_or(ErrorCode::NotFound)?;
-            if !matches!(mode, OpenMode::Read) {
-                return Err(ErrorCode::NoAccess);
-            }
-            Ok(bin_qid(node))
-        }
-
-        async fn read(&self, fid: Fid, offset: u64, count: u32) -> Result<Vec<u8>, ErrorCode> {
-            let fids = self.fids.lock().await;
-            let node = *fids.get(&fid).ok_or(ErrorCode::NotFound)?;
-            let bytes = match node {
-                BinNode::Root => BUILTIN_BIN_TOOLS.join("\n").into_bytes(),
-                BinNode::Tool(_) => Vec::new(),
-            };
-            let start = (offset as usize).min(bytes.len());
-            let end = bytes.len().min(start + count as usize);
-            Ok(bytes[start..end].to_vec())
-        }
-
-        async fn write(&self, _fid: Fid, _offset: u64, _data: &[u8]) -> Result<u32, ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn stat(&self, fid: Fid) -> Result<Stat, ErrorCode> {
-            let fids = self.fids.lock().await;
-            let node = *fids.get(&fid).ok_or(ErrorCode::NotFound)?;
-            Ok(Stat {
-                name: String::new(),
-                qid: bin_qid(node),
-                length: match node {
-                    BinNode::Root => BUILTIN_BIN_TOOLS.join("\n").len() as u64,
-                    BinNode::Tool(_) => 0,
-                },
-                executable: false,
-                writable: false,
-            })
-        }
-
-        async fn create(
-            &self,
-            _fid: Fid,
-            _newfid: Fid,
-            _name: &str,
-            _kind: FileKind,
-        ) -> Result<Qid, ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn remove(&self, _fid: Fid) -> Result<(), ErrorCode> {
-            Err(ErrorCode::Unsupported)
-        }
-
-        async fn clunk(&self, fid: Fid) -> Result<(), ErrorCode> {
-            if fid != Fid::ROOT {
-                self.fids.lock().await.remove(&fid);
-            }
-            Ok(())
-        }
-    }
-
-    fn bin_qid(node: BinNode) -> Qid {
-        match node {
-            BinNode::Root => Qid {
-                kind: FileKind::Dir,
-                version: 0,
-                path: 1,
-            },
-            BinNode::Tool(name) => Qid {
-                kind: FileKind::File,
-                version: 0,
-                path: 2 + BUILTIN_BIN_TOOLS
-                    .iter()
-                    .position(|tool| *tool == name)
-                    .unwrap_or(0) as u64,
-            },
-        }
-    }
-
-    struct JsonToolRunner;
-
-    #[async_trait]
-    impl ProcessRunner for JsonToolRunner {
-        async fn run(&self, invocation: ProcessInvocation) -> ProcessOutcome {
-            let Ok(resolved) = invocation.namespace.resolve(&invocation.exec.executable) else {
-                return ProcessOutcome::exited(
-                    127,
-                    br#"{"success":false,"error":"executable is not mounted"}"#.to_vec(),
-                );
-            };
-            let fid = Fid(70_000 + invocation.pid.0);
-            let reachable = resolved
-                .call(Request::Walk {
-                    fid: Fid::ROOT,
-                    newfid: fid,
-                    names: resolved.rel.clone(),
-                })
-                .await
-                .is_ok();
-            let _ = resolved.call(Request::Clunk { fid }).await;
-            if !reachable {
-                return ProcessOutcome::exited(
-                    127,
-                    br#"{"success":false,"error":"executable is not reachable"}"#.to_vec(),
-                );
-            }
-            let arguments = invocation
-                .exec
-                .args
-                .first()
-                .and_then(|raw| serde_json::from_str::<Value>(raw).ok())
-                .unwrap_or(Value::Null);
-            let tool_name = invocation
-                .exec
-                .executable
-                .rsplit('/')
-                .next()
-                .unwrap_or(invocation.exec.executable.as_str());
-            let mut output = json!({
-                "success": true,
-                "tool": tool_name,
-                "content": format!("from namespace {tool_name}"),
-                "arguments": arguments,
-            })
-            .to_string()
-            .into_bytes();
-            output.push(b'\n');
-            ProcessOutcome::exited(0, output)
-        }
-    }
-
-    struct RegistryToolRunner {
-        tools: ToolRegistry,
-    }
-
-    impl RegistryToolRunner {
-        fn new(tools: ToolRegistry) -> Self {
-            Self { tools }
-        }
-    }
-
-    #[async_trait]
-    impl ProcessRunner for RegistryToolRunner {
-        async fn run(&self, invocation: ProcessInvocation) -> ProcessOutcome {
-            if invocation
-                .namespace
-                .resolve(&invocation.exec.executable)
-                .is_err()
-            {
-                return ProcessOutcome::exited(127, b"executable is not mounted\n".to_vec());
-            }
-            let tool_name = invocation
-                .exec
-                .executable
-                .rsplit('/')
-                .next()
-                .unwrap_or(invocation.exec.executable.as_str());
-            let arguments = invocation
-                .exec
-                .args
-                .first()
-                .and_then(|raw| serde_json::from_str::<Value>(raw).ok())
-                .unwrap_or(Value::Null);
-
-            match self.tools.execute(tool_name, arguments).await {
-                Ok(output) => {
-                    let mut bytes = serde_json::to_vec(&output)
-                        .unwrap_or_else(|_| b"{\"success\":true}".to_vec());
-                    bytes.push(b'\n');
-                    ProcessOutcome::exited(0, bytes)
-                }
-                Err(err) => {
-                    let mut bytes = serde_json::to_vec(&json!({
-                        "success": false,
-                        "error": format!("{err:#}"),
-                    }))
-                    .unwrap_or_else(|_| b"{\"success\":false}".to_vec());
-                    bytes.push(b'\n');
-                    ProcessOutcome::exited(1, bytes)
-                }
-            }
-        }
-    }
-
-    struct CountingEffectTool {
-        name: &'static str,
-        capability: ToolCapability,
-        counter: Arc<AtomicUsize>,
-    }
-
-    impl Tool for CountingEffectTool {
-        fn name(&self) -> &str {
-            self.name
-        }
-
-        fn description(&self) -> &str {
-            "Counting side-effect tool used for dedupe tests"
-        }
-
-        fn parameters_schema(&self) -> Value {
-            json!({
-                "type": "object",
-                "properties": {
-                    "payload": {"type": "string"}
-                }
-            })
-        }
-
-        fn execute(&self, arguments: Value, _ctx: &ToolContext) -> ToolResult {
-            let counter = Arc::clone(&self.counter);
-            Box::pin(async move {
-                counter.fetch_add(1, Ordering::SeqCst);
-                Ok(json!({
-                    "ok": true,
-                    "payload": arguments
-                }))
-            })
-        }
-
-        fn capability(&self, _arguments: &Value) -> ToolCapability {
-            self.capability
-        }
-    }
-
-    fn create_test_state() -> RuntimeLoopState {
-        let config = Config::default();
-        let machine = AgentMachine::new();
-        let runtime_config = super::super::RuntimeConfig::default();
-        let mut namespace = Namespace::new();
-        namespace.mount(
-            "/agent/1",
-            InProcessTransport::new(Arc::new(AgentFs::new())),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(namespace)));
-
-        RuntimeLoopState {
-            machine,
-            current_submission_id: None,
-            environment: NamespaceRuntimeEnvironment::new(root, "/agent/1", "default"),
-            core_config: config,
-            runtime_config,
-            definition_persona_dirs: Vec::new(),
-            prompt_cache: crate::runtime::prompt_cache::PromptAssemblyCache::new(Vec::new()),
-            turn_state: TurnState::default(),
-        }
-    }
-
-    fn create_namespace_test_state_and_shell() -> (RuntimeLoopState, Shell) {
-        create_namespace_test_state_and_shell_with_bin(true)
-    }
-
-    fn create_namespace_test_state_and_shell_with_bin(
-        mount_bin: bool,
-    ) -> (RuntimeLoopState, Shell) {
-        create_namespace_test_state_and_shell_with_package(mount_bin, mount_bin)
-    }
-
-    fn create_namespace_test_state_and_shell_with_package(
-        mount_bin: bool,
-        mount_manifests: bool,
-    ) -> (RuntimeLoopState, Shell) {
-        let procfs = ProcFs::new().with_runner(Arc::new(JsonToolRunner));
-        let agentfs = Arc::new(AgentFs::new());
-        let bin = InProcessTransport::new(Arc::new(StaticBinFs::new()));
-        let mut manifests = Vec::new();
-
-        let mut child_namespace = Namespace::new();
-        child_namespace.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs.clone()),
-            Access::ReadWrite,
-        );
-        if mount_bin {
-            child_namespace.mount("/bin", bin.clone(), Access::ReadOnly);
-        }
-        if mount_manifests {
-            for tool_name in BUILTIN_BIN_TOOLS {
-                let manifest = crate::runtime::ToolPackageManifest {
-                    version: 1,
-                    name: tool_name.to_string(),
-                    description: format!("Test {tool_name} Tool"),
-                    parameters: json!({"type": "object"}),
-                    capability: ToolCapability::Read,
-                    capability_is_argument_dependent: false,
-                    timeout_secs: 30,
-                    execution: super::super::tool_packages::ToolExecutionHints {
-                        arguments: "json_first_arg".to_string(),
-                        result: "stdout_json".to_string(),
-                    },
-                };
-                let path = format!("/lib/exec/{tool_name}");
-                let transport = InProcessTransport::new(Arc::new(
-                    alan_ap::reference::MemFs::with_read_only_file(
-                        "manifest",
-                        serde_json::to_vec(&manifest).unwrap(),
-                    ),
-                ));
-                child_namespace.mount(&path, transport.clone(), Access::ReadOnly);
-                manifests.push((path, transport));
-            }
-        }
-
-        let spawner_procfs =
-            Arc::new(procfs.for_spawner(None, child_namespace, Credentials::user("root-agent")));
-        let mut root_namespace = Namespace::new();
-        root_namespace.mount(
-            "/proc",
-            InProcessTransport::new(spawner_procfs),
-            Access::ReadWrite,
-        );
-        root_namespace.mount(
-            "/agent/1",
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        if mount_bin {
-            root_namespace.mount("/bin", bin, Access::ReadOnly);
-        }
-        for (path, transport) in manifests {
-            root_namespace.mount(&path, transport, Access::ReadOnly);
-        }
-        let root = InProcessTransport::new(Arc::new(MountFs::new(root_namespace)));
-        let shell = Shell::new(root.clone());
-        let state = RuntimeLoopState {
-            machine: AgentMachine::new(),
-            current_submission_id: None,
-            environment: NamespaceRuntimeEnvironment::new(root, "/agent/1", "default"),
-            core_config: Config::default(),
-            runtime_config: super::super::RuntimeConfig::default(),
-            definition_persona_dirs: Vec::new(),
-            prompt_cache: crate::runtime::prompt_cache::PromptAssemblyCache::new(Vec::new()),
-            turn_state: TurnState::default(),
-        };
-        (state, shell)
-    }
-
-    fn reviewer_response(decision: &str) -> alan_llm::GenerationResponse {
-        alan_llm::GenerationResponse {
-            content: format!("{{\"decision\":\"{decision}\",\"rationale\":\"test\"}}"),
-            thinking: None,
-            thinking_signature: None,
-            redacted_thinking: Vec::new(),
-            tool_calls: Vec::new(),
-            usage: None,
-            finish_reason: None,
-            provider_response_id: None,
-            provider_response_status: None,
-            warnings: Vec::new(),
-        }
-    }
-
-    fn escalating_state_with_reviewer(
-        counter: &Arc<AtomicUsize>,
-        decision: &str,
-    ) -> RuntimeLoopState {
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("do the thing");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            // Unknown capability → autonomous policy escalates → reviewer route.
-            name: "do_thing",
-            capability: ToolCapability::Unknown,
-            counter: Arc::clone(counter),
-        });
-        create_test_state_with_machine_tools_and_provider(
-            machine,
-            tools,
-            alan_llm::MockLlmProvider::new().with_response(reviewer_response(decision)),
-        )
-    }
-
-    #[tokio::test]
-    async fn reviewer_allow_executes_escalated_tool() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut state = escalating_state_with_reviewer(&counter, "allow");
-        let _ = execute_single_tool_call(&mut state, "c1", "do_thing", json!({})).await;
-        assert_eq!(
-            counter.load(Ordering::SeqCst),
-            1,
-            "reviewer-approved escalation should execute"
-        );
-    }
-
-    #[tokio::test]
-    async fn reviewer_deny_blocks_escalated_tool_and_feeds_back() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut state = escalating_state_with_reviewer(&counter, "deny");
-        let (_, events) = execute_single_tool_call(&mut state, "c1", "do_thing", json!({})).await;
-        assert_eq!(
-            counter.load(Ordering::SeqCst),
-            0,
-            "reviewer-denied escalation must not execute"
-        );
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::Warning { message } if message.contains("auto-review denied")
-        )));
-    }
-
-    #[tokio::test]
-    async fn reviewer_repeated_denials_trip_breaker_to_human() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut state = escalating_state_with_reviewer(&counter, "deny");
-        // Third consecutive denial trips the circuit breaker and pauses for a human.
-        execute_single_tool_call(&mut state, "c1", "do_thing", json!({})).await;
-        execute_single_tool_call(&mut state, "c2", "do_thing", json!({})).await;
-        let (_, events) = execute_single_tool_call(&mut state, "c3", "do_thing", json!({})).await;
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::Yield {
-                kind: alan_agent_protocol::YieldKind::Confirmation,
-                ..
-            }
-        )));
-        assert_eq!(counter.load(Ordering::SeqCst), 0);
-    }
-
-    fn create_test_state_with_machine_and_tools(
-        machine: AgentMachine,
-        tools: ToolRegistry,
-    ) -> RuntimeLoopState {
-        create_test_state_with_machine_tools_and_provider(machine, tools, SimpleMockProvider)
-    }
-
-    fn create_test_state_with_machine_tools_and_provider<P: LlmProvider + 'static>(
-        machine: AgentMachine,
-        tools: ToolRegistry,
-        provider: P,
-    ) -> RuntimeLoopState {
-        create_test_state_with_machine_tools_provider_and_agent_path(
-            machine, tools, provider, "/agent/1",
-        )
-    }
-
-    fn create_test_state_with_machine_tools_provider_and_agent_path<P: LlmProvider + 'static>(
-        machine: AgentMachine,
-        mut tools: ToolRegistry,
-        provider: P,
-        agent_path: &str,
-    ) -> RuntimeLoopState {
-        let agentfs = Arc::new(AgentFs::new());
-        let llmfs = Arc::new(alan_llmfs::LlmFs::new());
-        llmfs.register_connection("default", Box::new(provider));
-
-        let mut process_namespace = Namespace::new();
-        process_namespace.mount(
-            agent_path,
-            InProcessTransport::new(agentfs),
-            Access::ReadWrite,
-        );
-        process_namespace.mount(
-            "/mnt/llm",
-            InProcessTransport::new(llmfs),
-            Access::ReadWrite,
-        );
-        for tool_name in tools.list_tools() {
-            process_namespace.mount(
-                &format!("/bin/{tool_name}"),
-                InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::new())),
-                Access::ReadOnly,
-            );
-            let tool = tools.get(tool_name).unwrap();
-            let manifest = crate::runtime::ToolPackageManifest::from_tool(
-                tool.as_ref(),
-                tools.execution_timeout_secs(tool_name).unwrap_or(30),
-            )
-            .unwrap();
-            process_namespace.mount(
-                &format!("/lib/exec/{tool_name}"),
-                InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::with_read_only_file(
-                    "manifest",
-                    serde_json::to_vec(&manifest).unwrap(),
-                ))),
-                Access::ReadOnly,
-            );
-        }
-
-        let launch_context = crate::ProcessLaunchContext::new(
-            process_namespace.clone(),
-            Credentials::user("test-agent"),
-            "/mnt/source",
-        )
-        .unwrap()
-        .with_host_mount(
-            crate::HostMountGrant::new("/mnt/source", "/tmp", Access::ReadWrite).unwrap(),
-        );
-        let binding = crate::tools::ToolExecutionBinding::from_launch_context(
-            &launch_context,
-            PathBuf::from("/tmp/alan-agent-engine-test-scratch"),
-        )
-        .unwrap();
-        tools.set_default_execution_binding(binding.clone());
-
-        let procfs = ProcFs::new().with_runner(Arc::new(RegistryToolRunner::new(tools.clone())));
-        let tool_runner = crate::tools::ToolProcessRunner::from_registry(&tools);
-        tool_runner.register_process_binding(alan_kernel::Pid(1), binding);
-        let spawner_procfs = Arc::new(procfs.for_spawner(
-            None,
-            process_namespace.clone(),
-            Credentials::user("root-agent"),
-        ));
-        process_namespace.mount(
-            "/proc",
-            InProcessTransport::new(spawner_procfs),
-            Access::ReadWrite,
-        );
-        let root = InProcessTransport::new(Arc::new(MountFs::new(process_namespace)));
-        let config = Config {
-            openai_responses_model: "mock-model".to_string(),
-            ..Default::default()
-        };
-        RuntimeLoopState {
-            machine,
-            current_submission_id: None,
-            environment: NamespaceRuntimeEnvironment::new(root, agent_path, "default")
-                .with_launch_context(launch_context)
-                .with_tool_process_context(alan_kernel::Pid(1), tool_runner),
-            core_config: config,
-            runtime_config: super::super::RuntimeConfig::default(),
-            definition_persona_dirs: Vec::new(),
-            prompt_cache: crate::runtime::prompt_cache::PromptAssemblyCache::new(Vec::new()),
-            turn_state: TurnState::default(),
-        }
-    }
-
-    async fn execute_single_tool_call(
-        state: &mut RuntimeLoopState,
-        call_id: &str,
-        tool_name: &str,
-        arguments: Value,
-    ) -> (ToolBatchOrchestratorOutcome, Vec<Event>) {
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-        let tool_calls = vec![NormalizedToolCall {
-            id: call_id.to_string(),
-            name: tool_name.to_string(),
-            arguments,
-        }];
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let mut events = Vec::new();
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let outcome = orchestrator
-            .orchestrate_tool_batch(state, &tool_calls, inputs, &mut emit)
-            .await
-            .expect("tool orchestration should succeed");
-        (outcome, events)
-    }
-
-    async fn read_shell_utf8(shell: &Shell, path: &str) -> String {
-        String::from_utf8(shell.cat(path).await.expect("read agent file")).expect("agent file utf8")
-    }
-
-    #[tokio::test]
-    async fn namespace_tool_call_spawns_bin_executable_and_records_action() {
-        let (mut state, shell) = create_namespace_test_state_and_shell();
-
-        let (outcome, events) = execute_single_tool_call(
-            &mut state,
-            "call-read",
-            "read_file",
-            json!({ "path": "sample.txt" }),
-        )
-        .await;
-
-        assert!(matches!(
-            outcome,
-            ToolBatchOrchestratorOutcome::ContinueTurnLoop {
-                refresh_context: false
-            }
-        ));
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::ToolCallCompleted {
-                id,
-                name,
-                success: Some(true),
-                ..
-            } if id == "call-read" && name.as_deref() == Some("read_file")
-        )));
-        let payload = state
-            .machine
-            .tool_payload_by_call_id("call-read")
-            .expect("tool response payload should be recorded on tape");
-        assert_eq!(payload["success"], json!(true));
-        assert_eq!(payload["content"], json!("from namespace read_file"));
-        assert_eq!(payload["arguments"], json!({ "path": "sample.txt" }));
-        assert_eq!(payload["exit_code"], json!(0));
-        assert_eq!(payload["process"], json!("/proc/1"));
-        assert_eq!(payload["action_id"], json!("a0"));
-
-        assert_eq!(
-            String::from_utf8(shell.cat("/proc/1/status").await.unwrap()).unwrap(),
-            "exited\n"
-        );
-        let action_output =
-            String::from_utf8(shell.cat("/agent/1/actions/a0/output").await.unwrap()).unwrap();
-        let action_payload: Value = serde_json::from_str(action_output.trim()).unwrap();
-        assert_eq!(action_payload["content"], json!("from namespace read_file"));
-        assert_eq!(
-            String::from_utf8(shell.cat("/agent/1/actions/a0/process").await.unwrap()).unwrap(),
-            "/proc/1"
-        );
-        assert_eq!(
-            String::from_utf8(shell.cat("/agent/1/actions/a0/result").await.unwrap()).unwrap(),
-            r#"{"exit_code":0}"#
-        );
-    }
-
-    #[tokio::test]
-    async fn long_tool_output_projects_to_resolvable_action_output_reference() {
-        let (mut state, shell) = create_namespace_test_state_and_shell();
-        let long_text = "x".repeat(crate::evidence::MAX_INLINE_EVIDENCE_BYTES + 1);
-
-        execute_single_tool_call(
-            &mut state,
-            "call-long",
-            "read_file",
-            json!({ "content": long_text }),
-        )
-        .await;
-
-        let payload = state
-            .machine
-            .tool_payload_by_call_id("call-long")
-            .expect("projected tool payload should be on tape");
-        assert_eq!(payload["type"], "evidence_projection");
-        assert_eq!(payload["reference"]["path"], "/agent/1/actions/a0/output");
-        assert_eq!(payload["truncation"]["full_content_recoverable"], true);
-        let full = read_shell_utf8(&shell, "/agent/1/actions/a0/output").await;
-        assert!(full.len() > crate::evidence::MAX_INLINE_EVIDENCE_BYTES);
-        assert!(full.contains(&"x".repeat(1024)));
-    }
-
-    #[tokio::test]
-    async fn short_tool_output_redacts_embedded_secrets_before_tape() {
-        let (mut state, _shell) = create_namespace_test_state_and_shell();
-
-        execute_single_tool_call(
-            &mut state,
-            "call-redacted-short",
-            "read_file",
-            json!({
-                "output": "api_key=short-secret\nAuthorization: Bearer short-token"
-            }),
-        )
-        .await;
-
-        let payload = state
-            .machine
-            .tool_payload_by_call_id("call-redacted-short")
-            .expect("inline tool payload should be on tape");
-        let serialized = serde_json::to_string(&payload).unwrap();
-        assert!(!serialized.contains("short-secret"));
-        assert!(!serialized.contains("short-token"));
-        assert!(serialized.contains("[REDACTED reason=secret_key]"));
-    }
-
-    #[tokio::test]
-    async fn redaction_expansion_still_projects_tool_payload_to_bounded_evidence() {
-        let (mut state, _shell) = create_namespace_test_state_and_shell();
-
-        execute_single_tool_call(
-            &mut state,
-            "call-redaction-expanded",
-            "read_file",
-            json!({
-                "output": "Bearer x ".repeat(1_000)
-            }),
-        )
-        .await;
-
-        let payload = state
-            .machine
-            .tool_payload_by_call_id("call-redaction-expanded")
-            .expect("expanded payload should be projected");
-        assert_eq!(payload["type"], "evidence_projection");
-        assert_eq!(payload["reference"]["path"], "/agent/1/actions/a0/output");
-        assert!(
-            payload["truncation"]["original_bytes"]
-                .as_u64()
-                .is_some_and(|bytes| bytes > crate::evidence::MAX_INLINE_EVIDENCE_BYTES as u64)
-        );
-    }
-
-    #[tokio::test]
-    async fn long_tool_output_uses_runtime_action_id_for_evidence_reference() {
-        let (mut state, _shell) = create_namespace_test_state_and_shell();
-
-        execute_single_tool_call(
-            &mut state,
-            "call-forged-action-id",
-            "read_file",
-            json!({
-                "action_id": "forged",
-                "content": "x".repeat(crate::evidence::MAX_INLINE_EVIDENCE_BYTES + 1)
-            }),
-        )
-        .await;
-
-        let payload = state
-            .machine
-            .tool_payload_by_call_id("call-forged-action-id")
-            .expect("projected tool payload should be on tape");
-        assert_eq!(payload["metadata"]["action_id"], "a0");
-        assert_eq!(payload["reference"]["path"], "/agent/1/actions/a0/output");
-    }
-
-    #[tokio::test]
-    async fn long_tool_output_without_action_path_uses_marked_inline_fallback() {
-        let state = create_test_state();
-        let payload = json!({
-            "content": "x".repeat(crate::evidence::MAX_INLINE_EVIDENCE_BYTES + 1)
-        });
-
-        let projected = tool_payload_for_tape(&state, &payload).await;
-
-        assert!(projected.get("reference").is_none());
-        assert_eq!(
-            projected["truncation"]["fallback_reason"],
-            "reference_unresolvable"
-        );
-        assert_eq!(projected["truncation"]["full_content_recoverable"], false);
-    }
-
-    #[tokio::test]
-    async fn durable_action_evidence_marks_secret_redaction_separately_from_truncation() {
-        let (mut state, shell) = create_namespace_test_state_and_shell();
-        execute_single_tool_call(
-            &mut state,
-            "call-redacted-long",
-            "read_file",
-            json!({
-                "authorization": "Bearer top-secret",
-                "content": "x".repeat(crate::evidence::MAX_INLINE_EVIDENCE_BYTES + 1)
-            }),
-        )
-        .await;
-
-        let payload = state
-            .machine
-            .tool_payload_by_call_id("call-redacted-long")
-            .unwrap();
-        assert!(!payload["preview"].as_str().unwrap().contains("top-secret"));
-        assert!(
-            payload["preview"]
-                .as_str()
-                .unwrap()
-                .contains("[REDACTED reason=secret_key]")
-        );
-        assert_eq!(payload["redactions"][0]["reason_class"], "secret_key");
-        assert_eq!(payload["truncation"]["full_content_recoverable"], true);
-        let full = read_shell_utf8(&shell, "/agent/1/actions/a0/output").await;
-        assert!(full.contains("[REDACTED reason=secret_key]"));
-        assert!(!full.contains("top-secret"));
-    }
-
-    #[tokio::test]
-    async fn namespace_tool_call_fails_when_bin_tool_is_withheld() {
-        let (mut state, shell) = create_namespace_test_state_and_shell_with_bin(false);
-
-        let (outcome, events) = execute_single_tool_call(
-            &mut state,
-            "call-read",
-            "read_file",
-            json!({ "path": "sample.txt" }),
-        )
-        .await;
-
-        assert!(matches!(
-            outcome,
-            ToolBatchOrchestratorOutcome::ContinueTurnLoop {
-                refresh_context: false
-            }
-        ));
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::ToolCallCompleted {
-                id,
-                name,
-                success: Some(false),
-                ..
-            } if id == "call-read" && name.as_deref() == Some("read_file")
-        )));
-        let payload = state
-            .machine
-            .tool_payload_by_call_id("call-read")
-            .expect("failed tool response payload should be recorded on tape");
-        assert_eq!(payload["success"], json!(false));
-        assert!(
-            payload["error"]
-                .as_str()
-                .is_some_and(|error| error.contains("not both mounted")),
-            "payload should explain the withheld executable: {payload}"
-        );
-        assert!(
-            shell.cat("/proc/1/status").await.is_err(),
-            "a withheld executable must not spawn"
-        );
-    }
-
-    #[tokio::test]
-    async fn namespace_tool_call_does_not_spawn_executable_without_manifest() {
-        let (mut state, shell) = create_namespace_test_state_and_shell_with_package(true, false);
-
-        let (outcome, events) = execute_single_tool_call(
-            &mut state,
-            "call-read",
-            "read_file",
-            json!({ "path": "sample.txt" }),
-        )
-        .await;
-
-        assert!(matches!(
-            outcome,
-            ToolBatchOrchestratorOutcome::ContinueTurnLoop {
-                refresh_context: false
-            }
-        ));
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::ToolCallCompleted {
-                id,
-                success: Some(false),
-                ..
-            } if id == "call-read"
-        )));
-        let payload = state
-            .machine
-            .tool_payload_by_call_id("call-read")
-            .expect("failed Tool response should be recorded on tape");
-        assert_eq!(payload["success"], json!(false));
-        assert!(
-            payload["error"]
-                .as_str()
-                .is_some_and(|error| error.contains("valid manifest"))
-        );
-        assert!(
-            shell.cat("/proc/1/status").await.is_err(),
-            "an executable without a valid Tool manifest must not spawn"
-        );
-    }
-
-    #[tokio::test]
-    async fn namespace_execution_target_spawns_every_builtin_bin_tool() {
-        let (state, shell) = create_namespace_test_state_and_shell();
-        let namespace = state.namespace_environment().clone();
-        let cancel = CancellationToken::new();
-
-        for (idx, tool_name) in BUILTIN_BIN_TOOLS.iter().enumerate() {
-            let payload = execute_tool_effect(
-                namespace.clone(),
-                tool_name,
-                json!({ "tool": tool_name, "call_index": idx }),
-                &cancel,
-                30,
-            )
-            .await
-            .expect("namespace tool effect should execute through /bin");
-            let pid = idx + 1;
-            let action_id = format!("a{idx}");
-
-            assert_eq!(payload["success"], json!(true));
-            assert_eq!(payload["tool"], json!(tool_name));
-            assert_eq!(
-                payload["content"],
-                json!(format!("from namespace {tool_name}"))
-            );
-            assert_eq!(
-                payload["arguments"],
-                json!({ "tool": tool_name, "call_index": idx })
-            );
-            assert_eq!(payload["exit_code"], json!(0));
-            assert_eq!(payload["process"], json!(format!("/proc/{pid}")));
-            assert_eq!(payload["action_id"], json!(action_id));
-
-            assert_eq!(
-                String::from_utf8(
-                    shell
-                        .cat(&format!("/agent/1/actions/{action_id}/name"))
-                        .await
-                        .unwrap()
-                )
-                .unwrap(),
-                *tool_name
-            );
-            assert_eq!(
-                String::from_utf8(
-                    shell
-                        .cat(&format!("/agent/1/actions/{action_id}/process"))
-                        .await
-                        .unwrap()
-                )
-                .unwrap(),
-                format!("/proc/{pid}")
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_tool_turn_orchestrator_new() {
-        let orchestrator = ToolTurnOrchestrator::new(Some(10), 4);
-        // Verify orchestrator was created with the correct settings
-        // Just test that it doesn't panic
-        let _ = orchestrator;
-    }
-
-    #[tokio::test]
-    async fn test_orchestrate_empty_tool_batch() {
-        let mut state = create_test_state();
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let tool_calls: Vec<NormalizedToolCall> = vec![];
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        assert!(result.is_ok());
-        match result.unwrap() {
-            ToolBatchOrchestratorOutcome::ContinueTurnLoop { refresh_context } => {
-                assert!(!refresh_context);
-            }
-            _ => panic!("Expected ContinueTurnLoop"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_handle_queued_steering_inputs_enforces_buffer_cap_for_follow_up() {
-        let mut state = create_test_state();
-        for idx in 0..MAX_BUFFERED_INBAND_USER_INPUTS {
-            state
-                .turn_state
-                .push_buffered_inband_submission(alan_agent_protocol::Submission::new(Op::Input {
-                    parts: vec![alan_agent_protocol::ContentPart::text(format!(
-                        "buffered-{idx}"
-                    ))],
-                    mode: InputMode::FollowUp,
-                }));
-        }
-        let broker = TurnInputBroker::default();
-        assert!(
-            broker
-                .push(alan_agent_protocol::Submission::new(Op::Input {
-                    parts: vec![alan_agent_protocol::ContentPart::text("overflow-follow-up")],
-                    mode: InputMode::FollowUp,
-                }))
-                .await
-        );
-
-        let mut events = Vec::new();
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let handled = handle_queued_steering_inputs(&mut state, &[], 0, Some(&broker), &mut emit)
-            .await
-            .unwrap();
-        assert!(!handled);
-        assert_eq!(
-            state.turn_state.buffered_inband_user_input_count(),
-            MAX_BUFFERED_INBAND_USER_INPUTS
-        );
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::Error { message, recoverable }
-                if *recoverable && message.contains("Too many queued in-turn user inputs")
-        )));
-    }
-
-    #[tokio::test]
-    async fn queued_steering_input_invalidates_earlier_active_plan() {
-        let mut state = create_test_state();
-        state
-            .turn_state
-            .begin_turn(state.machine.tape.messages().len());
-        state.machine.add_user_message("Initial task");
-        state.turn_state.set_plan_snapshot_at_message_count(
-            Some("Initial plan".to_string()),
-            Vec::new(),
-            state.machine.tape.messages().len(),
-        );
-        assert!(state.turn_state.plan_snapshot_is_from_active_turn());
-
-        let broker = TurnInputBroker::default();
-        assert!(
-            broker
-                .push(alan_agent_protocol::Submission::new(Op::Input {
-                    parts: vec![alan_agent_protocol::ContentPart::text(
-                        "Steer to the new task"
-                    )],
-                    mode: InputMode::Steer,
-                }))
-                .await
-        );
-        let mut emit = |_event: Event| async {};
-
-        let handled = handle_queued_steering_inputs(&mut state, &[], 0, Some(&broker), &mut emit)
-            .await
-            .unwrap();
-
-        assert!(handled);
-        assert!(!state.turn_state.plan_snapshot_is_from_active_turn());
-        assert_eq!(
-            state.machine.tape.messages().last().unwrap().text_content(),
-            "Steer to the new task"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_orchestrate_tool_batch_with_virtual_update_plan() {
-        let mut state = create_test_state();
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let tool_calls = vec![NormalizedToolCall {
-            id: "call_1".to_string(),
-            name: "update_plan".to_string(),
-            arguments: json!({
-                "explanation": "Test plan",
-                "items": [
-                    {"id": "1", "content": "Step 1", "status": "in_progress"}
-                ]
-            }),
-        }];
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        assert!(result.is_ok());
-        let has_update_plan_completion = events.iter().any(|event| {
-            matches!(
-                event,
-                Event::ToolCallCompleted {
-                    id,
-                    result_preview: Some(preview),
-                    ..
-                } if id == "call_1" && preview.contains("plan_updated")
-            )
-        });
-        assert!(
-            has_update_plan_completion,
-            "Expected update_plan ToolCallCompleted preview"
-        );
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::PlanUpdated { explanation, items }
-                if explanation.as_deref() == Some("Test plan")
-                    && items.len() == 1
-                    && items[0].content == "Step 1"
-        )));
-    }
-
-    #[tokio::test]
-    async fn test_orchestrate_tool_batch_with_virtual_confirmation() {
-        let mut state = create_test_state();
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let tool_calls = vec![NormalizedToolCall {
-            id: "call_1".to_string(),
-            name: "request_confirmation".to_string(),
-            arguments: json!({
-                "checkpoint_id": "chk_123",
-                "checkpoint_type": "test",
-                "summary": "Test confirmation",
-                "details": {"key": "value"}
-            }),
-        }];
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        assert!(result.is_ok());
-        match result.unwrap() {
-            ToolBatchOrchestratorOutcome::PauseTurn => {
-                // Expected
-            }
-            _ => panic!("Expected PauseTurn"),
-        }
-
-        // Check that Yield Confirmation event was emitted
-        let has_confirmation = events.iter().any(|e| {
-            matches!(
-                e,
-                Event::Yield {
-                    kind: alan_agent_protocol::YieldKind::Confirmation,
-                    ..
-                }
-            )
-        });
-        assert!(has_confirmation, "Expected Yield Confirmation event");
-    }
-
-    #[tokio::test]
-    async fn test_orchestrate_tool_batch_with_virtual_user_input() {
-        let mut state = create_test_state();
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let tool_calls = vec![NormalizedToolCall {
-            id: "call_1".to_string(),
-            name: "request_user_input".to_string(),
-            arguments: json!({
-                "title": "Test Input",
-                "prompt": "Enter something",
-                "questions": [
-                    {"id": "q1", "label": "Question 1", "prompt": "What?", "required": true}
-                ]
-            }),
-        }];
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        assert!(result.is_ok());
-        match result.unwrap() {
-            ToolBatchOrchestratorOutcome::PauseTurn => {
-                // Expected
-            }
-            _ => panic!("Expected PauseTurn"),
-        }
-
-        // Check that Yield event was emitted
-        let has_input_request = events.iter().any(|e| {
-            matches!(
-                e,
-                Event::Yield {
-                    kind: alan_agent_protocol::YieldKind::StructuredInput,
-                    ..
-                }
-            )
-        });
-        assert!(has_input_request, "Expected Yield StructuredInput event");
-    }
-
-    #[tokio::test]
-    async fn test_orchestrate_tool_batch_with_builtin_tool() {
-        let mut state = create_test_state();
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        // Test with read_file tool - requires sandbox setup, will likely fail but tests the path
-        let tool_calls = vec![NormalizedToolCall {
-            id: "call_1".to_string(),
-            name: "read_file".to_string(),
-            arguments: json!({"path": "test.txt"}),
-        }];
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        // Tool execution may fail due to sandbox restrictions, but orchestration should complete
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_replay_approved_tool_call() {
-        let mut state = create_test_state();
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let tool_call = NormalizedToolCall {
-            id: "call_1".to_string(),
-            name: "update_plan".to_string(),
-            arguments: json!({
-                "explanation": "Replay test",
-                "items": [{"id": "1", "content": "Step", "status": "completed"}]
-            }),
-        };
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = replay_approved_tool_call_with_cancel(
-            &mut state, &tool_call, None, None, inputs, &mut emit,
-        )
-        .await;
-
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_replay_approved_tool_batch() {
-        let mut state = create_test_state();
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let tool_calls = vec![NormalizedToolCall {
-            id: "call_1".to_string(),
-            name: "update_plan".to_string(),
-            arguments: json!({
-                "explanation": "Batch test",
-                "items": [{"id": "1", "content": "Step 1", "status": "completed"}]
-            }),
-        }];
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = replay_approved_tool_batch_with_cancel(
-            &mut state,
-            &tool_calls,
-            None,
-            None,
-            inputs,
-            &mut emit,
-        )
-        .await;
-
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_orchestrate_tool_batch_with_cancel() {
-        let mut state = create_test_state();
-        state.turn_state.begin_turn(0);
-        state
-            .turn_state
-            .set_turn_activity(crate::runtime::turn_state::TurnActivityState::Running);
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-
-        // Cancel immediately
-        cancel.cancel();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let tool_calls = vec![NormalizedToolCall {
-            id: "call_1".to_string(),
-            name: "read_file".to_string(),
-            arguments: json!({"path": "test.txt"}),
-        }];
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        // Should complete without panic even when cancelled
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_invalid_virtual_tool_ends_turn() {
-        let mut state = create_test_state();
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        // Invalid confirmation request - missing required summary
-        let tool_calls = vec![NormalizedToolCall {
-            id: "call_1".to_string(),
-            name: "request_confirmation".to_string(),
-            arguments: json!({
-                "details": {"reason": "missing_summary"}
-            }),
-        }];
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        assert!(result.is_ok());
-        // Invalid virtual tool should end turn
-        match result.unwrap() {
-            ToolBatchOrchestratorOutcome::EndTurn { .. } => {
-                // Check Error event was emitted
-                let has_error = events.iter().any(|e| matches!(e, Event::Error { .. }));
-                assert!(has_error, "Expected Error event for invalid virtual tool");
-            }
-            _ => panic!("Expected EndTurn for invalid virtual tool"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_multiple_tools_in_batch() {
-        let mut state = create_test_state();
-        let mut orchestrator = ToolTurnOrchestrator::new(None, 4);
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let tool_calls = vec![
-            NormalizedToolCall {
-                id: "call_1".to_string(),
-                name: "update_plan".to_string(),
-                arguments: json!({
-                    "explanation": "First",
-                    "items": [{"id": "1", "content": "Step 1", "status": "completed"}]
-                }),
-            },
-            NormalizedToolCall {
-                id: "call_2".to_string(),
-                name: "update_plan".to_string(),
-                arguments: json!({
-                    "explanation": "Second",
-                    "items": [{"id": "2", "content": "Step 2", "status": "completed"}]
-                }),
-            },
-        ];
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        assert!(result.is_ok());
-        // Should have two update_plan completion events.
-        let plan_updates: Vec<_> = events
-            .iter()
-            .filter(|event| {
-                matches!(
-                    event,
-                    Event::ToolCallCompleted {
-                        result_preview: Some(preview),
-                        ..
-                    } if preview.contains("plan_updated")
-                )
-            })
-            .collect();
-        assert_eq!(
-            plan_updates.len(),
-            2,
-            "Expected two update_plan completion events"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_side_effect_dedupe_survives_rollout_recovery_for_file_effects() {
-        let temp = tempfile::TempDir::new().unwrap();
-        let rollouts_dir = temp.path();
-        let counter = Arc::new(AtomicUsize::new(0));
-
-        let mut machine =
-            AgentMachine::new_with_recorder_in_dir("/proc/test", "mock", rollouts_dir)
-                .await
-                .unwrap();
-        machine.add_user_message("write file once");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            name: "write_file",
-            capability: ToolCapability::Write,
-            counter: Arc::clone(&counter),
-        });
-
-        let mut state = create_test_state_with_machine_and_tools(machine, tools);
-        assert_eq!(state.process_path(), "/proc/1");
-        let (_, first_events) = execute_single_tool_call(
-            &mut state,
-            "call-file-1",
-            "write_file",
-            json!({"path": "notes.txt", "payload": "hello"}),
-        )
-        .await;
-        assert!(
-            first_events
-                .iter()
-                .any(|event| matches!(event, Event::ToolCallCompleted { .. }))
-        );
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-
-        tokio::time::sleep(std::time::Duration::from_millis(80)).await;
-        let rollout_path = state
-            .machine
-            .recorder
-            .as_ref()
-            .expect("recorder should exist")
-            .path()
-            .clone();
-
-        let recovered_machine =
-            AgentMachine::load_from_rollout_in_dir(&rollout_path, "/proc/2", "mock", rollouts_dir)
-                .await
-                .unwrap();
-        let mut recovered_tools = ToolRegistry::new();
-        recovered_tools.register(CountingEffectTool {
-            name: "write_file",
-            capability: ToolCapability::Write,
-            counter: Arc::clone(&counter),
-        });
-        let mut recovered_state = create_test_state_with_machine_tools_provider_and_agent_path(
-            recovered_machine,
-            recovered_tools,
-            SimpleMockProvider,
-            "/agent/2",
-        );
-        assert_eq!(recovered_state.process_path(), "/proc/2");
-        let replay_arguments = json!({"path": "notes.txt", "payload": "hello"});
-        let replay_identity = build_effect_identity(
-            &recovered_state.machine,
-            "write_file",
-            &replay_arguments,
-            EffectCategory::File,
-        );
-        let _ = execute_single_tool_call(
-            &mut recovered_state,
-            "call-file-2",
-            "write_file",
-            replay_arguments,
-        )
-        .await;
-
-        assert_eq!(
-            counter.load(Ordering::SeqCst),
-            1,
-            "dedupe after recovery should skip physical execution"
-        );
-        assert_eq!(
-            recovered_state
-                .machine
-                .tool_payload_by_call_id("call-file-2")
-                .expect("replayed tool payload should exist"),
-            recovered_state
-                .machine
-                .tool_payload_by_call_id("call-file-1")
-                .expect("original tool payload should exist"),
-            "dedupe replay should preserve original tool payload"
-        );
-        let replayed_effect = recovered_state
-            .machine
-            .effect_by_idempotency_key(&replay_identity.idempotency_key)
-            .expect("dedupe replay effect should exist");
-        assert_eq!(replayed_effect.process_path, "/proc/2");
-        assert!(replayed_effect.dedupe_hit);
-    }
-
-    #[tokio::test]
-    async fn test_side_effect_dedupe_for_network_effects() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("call api once");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            name: "bash",
-            capability: ToolCapability::Network,
-            counter: Arc::clone(&counter),
-        });
-        let mut state = create_test_state_with_machine_and_tools(machine, tools);
-        // Exercise effect dedupe independent of the locked auto-approve posture
-        // (which would otherwise escalate the network call).
-        state.runtime_config.policy_engine = crate::policy::PolicyEngine::allow_all();
-        let arguments = json!({
-            "command": "curl https://example.com",
-            "output": "api_key=embedded-secret",
-            "headers": {
-                "authorization": "Bearer secret-token"
-            }
-        });
-        let identity =
-            build_effect_identity(&state.machine, "bash", &arguments, EffectCategory::Network);
-
-        let _ = execute_single_tool_call(&mut state, "call-net-1", "bash", arguments.clone()).await;
-        let _ = execute_single_tool_call(&mut state, "call-net-2", "bash", arguments).await;
-
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-        let replayed_payload = state
-            .machine
-            .tool_payload_by_call_id("call-net-2")
-            .expect("replayed tool payload should exist");
-        assert_eq!(
-            replayed_payload,
-            state
-                .machine
-                .tool_payload_by_call_id("call-net-1")
-                .expect("original tool payload should exist"),
-            "dedupe replay should preserve original network-tool payload"
-        );
-        assert_eq!(
-            replayed_payload["payload"]["headers"]["authorization"],
-            json!("[REDACTED reason=secret_key]"),
-            "dedupe replay should preserve the redacted tape payload"
-        );
-        let effect = state
-            .machine
-            .effect_by_idempotency_key(&identity.idempotency_key)
-            .expect("effect record should exist");
-        assert_eq!(
-            effect
-                .result_payload
-                .as_ref()
-                .and_then(|payload| payload.get("payload"))
-                .and_then(|payload| payload.get("headers"))
-                .and_then(|headers| headers.get("authorization")),
-            Some(&json!("[REDACTED reason=secret_key]")),
-            "durable effect payloads should stay redacted for persistence"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_effect_record_and_tape_payload_are_both_redacted() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("call api with auth");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            name: "bash",
-            capability: ToolCapability::Network,
-            counter: Arc::clone(&counter),
-        });
-        let mut state = create_test_state_with_machine_and_tools(machine, tools);
-        // Exercise effect recording independent of the locked auto-approve
-        // posture (which would otherwise escalate the network call).
-        state.runtime_config.policy_engine = crate::policy::PolicyEngine::allow_all();
-        let arguments = json!({
-            "command": "curl https://example.com",
-            "output": "api_key=embedded-secret",
-            "headers": {
-                "authorization": "Bearer secret-token"
-            }
-        });
-        let identity =
-            build_effect_identity(&state.machine, "bash", &arguments, EffectCategory::Network);
-
-        let _ = execute_single_tool_call(&mut state, "call-net-secret", "bash", arguments).await;
-
-        let effect = state
-            .machine
-            .effect_by_idempotency_key(&identity.idempotency_key)
-            .expect("effect record should exist");
-        assert_eq!(
-            effect
-                .result_payload
-                .as_ref()
-                .and_then(|payload| payload.get("payload"))
-                .and_then(|payload| payload.get("headers"))
-                .and_then(|headers| headers.get("authorization")),
-            Some(&json!("[REDACTED reason=secret_key]"))
-        );
-        assert_eq!(
-            effect
-                .result_payload
-                .as_ref()
-                .and_then(|payload| payload.get("payload"))
-                .and_then(|payload| payload.get("output")),
-            Some(&json!("api_key= [REDACTED reason=secret_key]"))
-        );
-
-        let tape_payload = state
-            .machine
-            .tool_payload_by_call_id("call-net-secret")
-            .expect("tool payload should exist on tape");
-        assert_eq!(
-            tape_payload["payload"]["headers"]["authorization"],
-            json!("[REDACTED reason=secret_key]")
-        );
-        assert_eq!(
-            tape_payload["payload"]["output"],
-            json!("api_key= [REDACTED reason=secret_key]")
-        );
-    }
-
-    #[tokio::test]
-    async fn test_side_effect_dedupe_for_process_effects() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("run command once");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            name: "bash",
-            capability: ToolCapability::Write,
-            counter: Arc::clone(&counter),
-        });
-        let mut state = create_test_state_with_machine_and_tools(machine, tools);
-
-        let _ = execute_single_tool_call(
-            &mut state,
-            "call-proc-1",
-            "bash",
-            json!({"command": "touch hello.txt"}),
-        )
-        .await;
-        let _ = execute_single_tool_call(
-            &mut state,
-            "call-proc-2",
-            "bash",
-            json!({"command": "touch hello.txt"}),
-        )
-        .await;
-
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-        assert_eq!(
-            state
-                .machine
-                .tool_payload_by_call_id("call-proc-2")
-                .expect("replayed tool payload should exist"),
-            state
-                .machine
-                .tool_payload_by_call_id("call-proc-1")
-                .expect("original tool payload should exist"),
-            "dedupe replay should preserve original process-tool payload"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_unknown_effect_status_escalates_without_execution() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("write file with safety");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            name: "write_file",
-            capability: ToolCapability::Write,
-            counter: Arc::clone(&counter),
-        });
-        let mut state = create_test_state_with_machine_and_tools(machine, tools);
-        let arguments = json!({"path": "notes.txt", "payload": "hello"});
-        let identity = build_effect_identity(
-            &state.machine,
-            "write_file",
-            &arguments,
-            EffectCategory::File,
-        );
-        state.machine.record_effect(crate::rollout::EffectRecord {
-            effect_id: "ef-unknown".to_string(),
-            process_path: state.process_path(),
-            tool_call_id: "call-prev".to_string(),
-            idempotency_key: identity.idempotency_key.clone(),
-            effect_type: "file".to_string(),
-            request_fingerprint: identity.request_fingerprint.clone(),
-            result_digest: None,
-            result_payload: None,
-            status: crate::rollout::EffectStatus::Unknown,
-            applied_at: None,
-            reason: Some("crash during prior execution".to_string()),
-            dedupe_hit: false,
-            timestamp: chrono::Utc::now().to_rfc3339(),
-        });
-
-        let (outcome, events) =
-            execute_single_tool_call(&mut state, "call-new", "write_file", arguments).await;
-        assert!(matches!(outcome, ToolBatchOrchestratorOutcome::PauseTurn));
-        assert_eq!(
-            counter.load(Ordering::SeqCst),
-            0,
-            "unknown effect status should not execute without confirmation"
-        );
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::Yield {
-                kind: alan_agent_protocol::YieldKind::Confirmation,
-                ..
-            }
-        )));
-    }
-
-    #[tokio::test]
-    async fn test_replay_approved_unknown_effect_executes_tool_once() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("write file with approval");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            name: "write_file",
-            capability: ToolCapability::Write,
-            counter: Arc::clone(&counter),
-        });
-        let mut state = create_test_state_with_machine_and_tools(machine, tools);
-        let arguments = json!({"path": "notes.txt", "payload": "hello"});
-        let identity = build_effect_identity(
-            &state.machine,
-            "write_file",
-            &arguments,
-            EffectCategory::File,
-        );
-        state.machine.record_effect(crate::rollout::EffectRecord {
-            effect_id: "ef-unknown".to_string(),
-            process_path: state.process_path(),
-            tool_call_id: "call-prev".to_string(),
-            idempotency_key: identity.idempotency_key.clone(),
-            effect_type: "file".to_string(),
-            request_fingerprint: identity.request_fingerprint.clone(),
-            result_digest: None,
-            result_payload: None,
-            status: crate::rollout::EffectStatus::Unknown,
-            applied_at: None,
-            reason: Some("crash during prior execution".to_string()),
-            dedupe_hit: false,
-            timestamp: chrono::Utc::now().to_rfc3339(),
-        });
-
-        let cancel = CancellationToken::new();
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-        let tool_call = NormalizedToolCall {
-            id: "call-new".to_string(),
-            name: "write_file".to_string(),
-            arguments,
-        };
-        let mut events = Vec::new();
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let outcome = replay_approved_tool_call_with_cancel(
-            &mut state,
-            &tool_call,
-            Some(tool_call.id.as_str()),
-            None,
-            inputs,
-            &mut emit,
-        )
-        .await
-        .expect("approved replay should run");
-        assert!(matches!(
-            outcome,
-            ToolBatchOrchestratorOutcome::ContinueTurnLoop { .. }
-        ));
-        assert_eq!(
-            counter.load(Ordering::SeqCst),
-            1,
-            "approved replay should execute once"
-        );
-        assert!(
-            !events.iter().any(|event| matches!(
-                event,
-                Event::Yield {
-                    kind: alan_agent_protocol::YieldKind::Confirmation,
-                    ..
-                }
-            )),
-            "approved replay should not emit a second confirmation yield"
-        );
-
-        let restored = state
-            .machine
-            .effect_by_idempotency_key(&identity.idempotency_key)
-            .expect("updated effect record should exist");
-        assert_eq!(restored.status, crate::rollout::EffectStatus::Applied);
-    }
-
-    #[tokio::test]
-    async fn test_replay_approved_tool_escalation_executes_tool_once() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("run unknown tool with approval");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            name: "unknown_effect_tool",
-            capability: ToolCapability::Unknown,
-            counter: Arc::clone(&counter),
-        });
-        let mut state = create_test_state_with_machine_and_tools(machine, tools);
-
-        let cancel = CancellationToken::new();
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-        let tool_call = NormalizedToolCall {
-            id: "call-approved".to_string(),
-            name: "unknown_effect_tool".to_string(),
-            arguments: json!({"payload": "hello"}),
-        };
-        let mut events = Vec::new();
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let outcome = replay_approved_tool_call_with_cancel(
-            &mut state,
-            &tool_call,
-            None,
-            Some(tool_call.id.as_str()),
-            inputs,
-            &mut emit,
-        )
-        .await
-        .expect("approved tool escalation replay should run");
-        assert!(matches!(
-            outcome,
-            ToolBatchOrchestratorOutcome::ContinueTurnLoop { .. }
-        ));
-        assert_eq!(
-            counter.load(Ordering::SeqCst),
-            1,
-            "approved tool escalation replay should execute once"
-        );
-        assert!(
-            !events.iter().any(|event| matches!(
-                event,
-                Event::Yield {
-                    kind: alan_agent_protocol::YieldKind::Confirmation,
-                    ..
-                }
-            )),
-            "approved tool escalation replay should not emit a second confirmation yield"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_replay_approved_batch_bypasses_unknown_only_for_first_tool_call() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("write file with batch replay");
-        let mut tools = ToolRegistry::new();
-        tools.register(CountingEffectTool {
-            name: "write_file",
-            capability: ToolCapability::Write,
-            counter: Arc::clone(&counter),
-        });
-        let mut state = create_test_state_with_machine_and_tools(machine, tools);
-        let arguments_first = json!({"path": "notes-1.txt", "payload": "hello"});
-        let arguments_second = json!({"path": "notes-2.txt", "payload": "world"});
-        let identity_first = build_effect_identity(
-            &state.machine,
-            "write_file",
-            &arguments_first,
-            EffectCategory::File,
-        );
-        let identity_second = build_effect_identity(
-            &state.machine,
-            "write_file",
-            &arguments_second,
-            EffectCategory::File,
-        );
-        state.machine.record_effect(crate::rollout::EffectRecord {
-            effect_id: "ef-unknown-1".to_string(),
-            process_path: state.process_path(),
-            tool_call_id: "call-prev-1".to_string(),
-            idempotency_key: identity_first.idempotency_key.clone(),
-            effect_type: "file".to_string(),
-            request_fingerprint: identity_first.request_fingerprint.clone(),
-            result_digest: None,
-            result_payload: None,
-            status: crate::rollout::EffectStatus::Unknown,
-            applied_at: None,
-            reason: Some("crash during prior execution".to_string()),
-            dedupe_hit: false,
-            timestamp: chrono::Utc::now().to_rfc3339(),
-        });
-        state.machine.record_effect(crate::rollout::EffectRecord {
-            effect_id: "ef-unknown-2".to_string(),
-            process_path: state.process_path(),
-            tool_call_id: "call-prev-2".to_string(),
-            idempotency_key: identity_second.idempotency_key.clone(),
-            effect_type: "file".to_string(),
-            request_fingerprint: identity_second.request_fingerprint.clone(),
-            result_digest: None,
-            result_payload: None,
-            status: crate::rollout::EffectStatus::Unknown,
-            applied_at: None,
-            reason: Some("crash during prior execution".to_string()),
-            dedupe_hit: false,
-            timestamp: chrono::Utc::now().to_rfc3339(),
-        });
-
-        let tool_calls = vec![
-            NormalizedToolCall {
-                id: "call-dup".to_string(),
-                name: "write_file".to_string(),
-                arguments: arguments_first,
-            },
-            NormalizedToolCall {
-                id: "call-dup".to_string(),
-                name: "write_file".to_string(),
-                arguments: arguments_second,
-            },
-        ];
-        let cancel = CancellationToken::new();
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-        let mut events = Vec::new();
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let outcome = replay_approved_tool_batch_with_cancel(
-            &mut state,
-            &tool_calls,
-            Some("call-dup"),
-            None,
-            inputs,
-            &mut emit,
-        )
-        .await
-        .expect("approved replay batch should run");
-        assert!(matches!(outcome, ToolBatchOrchestratorOutcome::PauseTurn));
-        assert_eq!(
-            counter.load(Ordering::SeqCst),
-            1,
-            "only the approved call should bypass unknown-effect escalation"
-        );
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::Yield {
-                kind: alan_agent_protocol::YieldKind::Confirmation,
-                payload,
-                ..
-            } if payload["details"]["replay_tool_call"]["call_id"] == "call-dup"
-        )));
-    }
-
-    #[tokio::test]
-    async fn test_tool_loop_guard_triggers() {
-        let mut state = create_test_state();
-        // Set max loops to a small number
-        let mut orchestrator = ToolTurnOrchestrator::new(Some(2), 4);
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        // Create many tool calls that will exceed the loop limit
-        let mut tool_calls = vec![];
-        for i in 0..3 {
-            tool_calls.push(NormalizedToolCall {
-                id: format!("call_{}", i),
-                name: "update_plan".to_string(),
-                arguments: json!({
-                    "explanation": format!("Step {}", i),
-                    "items": [{"id": i.to_string(), "content": "Step", "status": "completed"}]
-                }),
-            });
-        }
-
-        let inputs = ToolOrchestratorInputs {
-            cancel: &cancel,
-            steering_broker: None,
-        };
-
-        let result = orchestrator
-            .orchestrate_tool_batch(&mut state, &tool_calls, inputs, &mut emit)
-            .await;
-
-        assert!(result.is_ok());
-        // After max loops, should end turn
-        match result.unwrap() {
-            ToolBatchOrchestratorOutcome::EndTurn { .. } => {
-                // Expected
-            }
-            _ => {
-                // Note: Depending on implementation, might continue or end
-                // Just verify no panic occurred
-            }
-        }
-    }
-}
+mod tests;
