+++ crates/agent-engine/src/runtime/tool_orchestrator/tests.rs
@@ -1,0 +1,3 @@
+include!("tests/support_and_namespace_contract.inc.rs");
+include!("tests/namespace_and_batch_contract.inc.rs");
+include!("tests/replay_and_effect_contract.inc.rs");
